w5_bruin/my-taxi-pipeline/pipeline/assets/ingestion/requirements.txt
pandas==2.2.0
requests==2.31.0
pyarrow==23.0.0
python-dateutil==2.9.0.post0
